Fix MSVC warning

+++ source/app/ui/document.cpp
@@ -316,7 +316,7 @@ QString Document::pluginQmlPath() const
 QStringList Document::bookmarks() const
 {
     QStringList list;
-    list.reserve(_bookmarks.size());
+    list.reserve(static_cast<int>(_bookmarks.size()));
 
     for(const auto& name : u::keysFor(_bookmarks))
         list.append(name);
